requirements.txt
Django>=2<=3
django-rest-framework
psycopg2>=2.7,<3.0